Refactor spotify-api login url
